Allow admins to set options for games
The intention here is to allow admins to create a game and set options
on it, so that initially we can choose to select quotes from Langflow or
an Astra DB collection.

To do so, we do the following:

* Create an /admin route with a similar setup to the home route
* Add options to the game creation form for admins
* Store the game options in the PartyServer
* Creates a barebones Astra DB client (@datastax/astra-db-ts doesn't
seem to run on Cloudflare with the settings we have currently)
* Uses the game options to choose between Langflow and Astra DB to load
questions

This works, so far, and opens the opportunity for other game settings
(like number of questions in a game) to be set by admins. This also does
some work towards storing other data in Astra DB, like player email
addresses or final game states.